refactor(inference): move test-only q8 helpers
